job-replaceability-analyser/requirements.txt
# Job Replaceability Analyser — Python dependencies
# Pin to exact versions for reproducible installs in production.
# Run: pip install -r requirements.txt

# --- Web framework ---
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1

# --- Configuration ---
pyyaml==6.0.1
python-dotenv==1.0.1

# --- HTTP client (for internal service calls) ---
httpx==0.27.0

# --- Reporting ---
jinja2==3.1.4           # PDF/Markdown report templating
weasyprint==62.1        # PDF generation from HTML templates

# --- Data & analysis ---
pandas==2.2.2           # Role benchmark data processing
numpy==1.26.4

# --- Testing ---
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0

# --- Development tools ---
ruff==0.4.4             # Linting + formatting
mypy==1.10.0            # Type checking
pre-commit==3.7.1
